# feat: speed up the new parse_advertisement_data_tuple function