--- a/src/bluetooth_data_tools/gap.pxd
+++ b/src/bluetooth_data_tools/gap.pxd
@@ -37,6 +37,8 @@ cdef cython.uint TYPE_TX_POWER_LEVEL
 
 cpdef parse_advertisement_data(object data)
 
+cpdef parse_advertisement_data_tuple(cython.tuple data)
+
 @cython.locals(
     gap_data=cython.bytes,
     gap_value=cython.bytes,