--- a/src/bluetooth_data_tools/gap.py
+++ b/src/bluetooth_data_tools/gap.py
@@ -179,7 +179,11 @@ def _parse_advertisement_data_tuple(
     return _uncached_parse_advertisement_data(data)
 
 
-parse_advertisement_data_tuple = _parse_advertisement_data_tuple
+def parse_advertisement_data_tuple(
+    data: Tuple[bytes, ...],
+) -> BLEGAPAdvertisementTupleType:
+    """Parse a tuple of raw advertisement data and return a tuple of BLEGAPAdvertisementTupleType."""
+    return _parse_advertisement_data_tuple(data)
 
 
 def _uncached_parse_advertisement_data(