… bench/parse_gap_tuple.py