This is technically a breaking change to only allow incoming data to be a tuple, but since its new (less than 2 weeks old) and only in HA dev, and likely only HA is using it, I did not bump the major version.

Before

Parsing 100000 bluetooth messages took 0.047129874998063315 seconds
Parsing 100000 bluetooth messages took 0.04788412500056438 seconds
Parsing 100000 bluetooth messages took 0.048007541001425125 seconds

After

Parsing 100000 bluetooth messages took 0.0215528330008965 seconds
Parsing 100000 bluetooth messages took 0.022052750002330868 seconds
Parsing 100000 bluetooth messages took 0.02225575000193203 seconds